refactor: split completion-generators into per-shell files, remove re-exports
- Split 215-line completion-generators.ts into completion-bash.ts,
  completion-zsh.ts, completion-fish.ts (one function per file)
- Remove re-exports of generator functions from commands/completion.ts
- Update test imports to pull generators directly from utils/ modules
- Delete completion-generators.ts

Fixes PR #175 review findings (200-line limit, no re-exports from commands/).
