+++ src/utils/completion-fish.ts
@@ -1,150 +1,3 @@
-const COMMANDS = "init install generate check snapshot status report hook completion";
-
-export function generateBashCompletion(): string {
-  return `# bash completion for ai-guardrails
-# Source this file or add to ~/.bash_completion.d/
-_ai_guardrails() {
-  local cur prev
-  cur="\${COMP_WORDS[COMP_CWORD]}"
-  prev="\${COMP_WORDS[COMP_CWORD-1]}"
-
-  if [ "$COMP_CWORD" -eq 1 ]; then
-    COMPREPLY=($(compgen -W "${COMMANDS}" -- "$cur"))
-    return
-  fi
-
-  case "\${COMP_WORDS[1]}" in
-    init)
-      COMPREPLY=($(compgen -W "--yes --profile --force --upgrade --interactive --no-hooks --no-ci --no-agent-rules --config-strategy --project-dir" -- "$cur"))
-      ;;
-    install)
-      COMPREPLY=($(compgen -W "--upgrade --project-dir" -- "$cur"))
-      ;;
-    generate)
-      COMPREPLY=($(compgen -W "--check --project-dir" -- "$cur"))
-      ;;
-    check)
-      COMPREPLY=($(compgen -W "--format --baseline --strict --project-dir" -- "$cur"))
-      ;;
-    snapshot)
-      COMPREPLY=($(compgen -W "--baseline --project-dir" -- "$cur"))
-      ;;
-    status|report)
-      COMPREPLY=($(compgen -W "--project-dir" -- "$cur"))
-      ;;
-    hook)
-      COMPREPLY=($(compgen -W "dangerous-cmd protect-configs protect-reads suppress-comments" -- "$cur"))
-      ;;
-    completion)
-      COMPREPLY=($(compgen -W "bash zsh fish" -- "$cur"))
-      ;;
-    *)
-      COMPREPLY=($(compgen -W "--project-dir --quiet --no-color --version --help" -- "$cur"))
-      ;;
-  esac
-}
-complete -F _ai_guardrails ai-guardrails
-`;
-}
-
-export function generateZshCompletion(): string {
-  return `#compdef ai-guardrails
-_ai_guardrails() {
-  local -a commands
-  commands=(
-    'init:Per-project setup'
-    'install:One-time machine setup'
-    'generate:Regenerate all managed config files'
-    'check:Hold-the-line enforcement'
-    'snapshot:Capture current lint state as baseline'
-    'status:Project health dashboard'
-    'report:Show recent check run history'
-    'hook:Internal hook dispatcher'
-    'completion:Generate shell completion script'
-  )
-
-  local -a global_opts
-  global_opts=(
-    '--project-dir[Override working directory]:dir:_files -/'
-    '--quiet[Suppress info/success output]'
-    '--no-color[Disable ANSI color output]'
-    '--version[Print version]'
-    '--help[Show help]'
-  )
-
-  _arguments -C \\
-    $global_opts \\
-    '1: :->command' \\
-    '*: :->args'
-
-  case $state in
-    command)
-      _describe 'command' commands
-      ;;
-    args)
-      case \${words[2]} in
-        init)
-          _arguments \\
-            '--yes[Accept all defaults]' \\
-            '--profile[Set profile]:profile:(strict standard minimal)' \\
-            '--force[Overwrite existing managed files]' \\
-            '--upgrade[Refresh generated files]' \\
-            '--interactive[Prompt for each optional step]' \\
-            '--no-hooks[Skip lefthook install]' \\
-            '--no-ci[Skip CI workflow generation]' \\
-            '--no-agent-rules[Skip AGENTS.md and IDE rule files]' \\
-            '--config-strategy[Config handling strategy]:strategy:(merge replace skip)' \\
-            '--project-dir[Override working directory]:dir:_files -/'
-          ;;
-        install)
-          _arguments \\
-            '--upgrade[Overwrite existing machine config]' \\
-            '--project-dir[Override working directory]:dir:_files -/'
-          ;;
-        generate)
-          _arguments \\
-            '--check[Verify files are up-to-date]' \\
-            '--project-dir[Override working directory]:dir:_files -/'
-          ;;
-        check)
-          _arguments \\
-            '--format[Output format]:format:(text sarif)' \\
-            '--baseline[Custom baseline path]:file:_files' \\
-            '--strict[Ignore baseline]' \\
-            '--project-dir[Override working directory]:dir:_files -/'
-          ;;
-        snapshot)
-          _arguments \\
-            '--baseline[Custom output path]:file:_files' \\
-            '--project-dir[Override working directory]:dir:_files -/'
-          ;;
-        status|report)
-          _arguments \\
-            '--project-dir[Override working directory]:dir:_files -/'
-          ;;
-        hook)
-          local -a hooks
-          hooks=(
-            'dangerous-cmd:Check for dangerous shell commands'
-            'protect-configs:Protect managed config files'
-            'protect-reads:Protect sensitive file reads'
-            'suppress-comments:Remove AI comment markers'
-          )
-          _describe 'hook' hooks
-          ;;
-        completion)
-          local -a shells
-          shells=('bash:Bash completion' 'zsh:Zsh completion' 'fish:Fish completion')
-          _describe 'shell' shells
-          ;;
-      esac
-      ;;
-  esac
-}
-_ai_guardrails
-`;
-}
-
 export function generateFishCompletion(): string {
   return `# fish completion for ai-guardrails
 complete -c ai-guardrails -f
